+++ bauh/gems/flatpak/model.py
@@ -50,7 +50,7 @@ class FlatpakApplication(SoftwarePackage):
         return self.get_default_icon_path()
 
     def is_application(self):
-        return not self.runtime
+        return not self.runtime or not self.partial
 
     def get_disk_cache_path(self):
         return super(FlatpakApplication, self).get_disk_cache_path() + '/installed/' + self.id
@@ -77,9 +77,20 @@ class FlatpakApplication(SoftwarePackage):
         return self.origin
 
     def gen_partial(self, partial_id: str) -> "FlatpakApplication":
-        partial = copy.deepcopy(self)
+        partial = FlatpakApplication()
+        partial.partial = True
         partial.id = partial_id
         partial.base_id = self.id
+        partial.installation = self.installation
+        partial.origin = self.origin
+        partial.branch = self.branch
+        partial.i18n = self.i18n
+        partial.arch = self.arch
+        partial.name = self.name
+        partial.version = self.version
+        partial.latest_version = self.latest_version
+        partial.installed = self.installed
+        partial.runtime = True
 
         if self.ref:
             partial.base_ref = self.ref
@@ -87,7 +98,6 @@ class FlatpakApplication(SoftwarePackage):
             partial.status = PackageStatus.READY
             partial.name += ' ( {} )'.format(partial_id.split('.')[-1])
 
-        partial.partial = True
         return partial
 
     def get_name_tooltip(self) -> str: